Improve bottleneck configuration
Fixes #9

It's no longer possible to accidentally limit the max concurrent jobs to zero, which would prevent any perforce commands from running.

Default settings now limit to 10 max concurrent jobs, while high water is set to 150 as a failsafe.

Descriptions have been updated to clarify documentation - e.g. setting a low value for high water seems to be dangerous - in that commands will be dropped without explanation.

The high water is more of a failsafe in case the extension goes crazy, so ideally it should not be required. Seems most likely to occur in large perforce workspaces when syncing, with edit on modified / save enabled?