Performance optimisations and code tidy-up
Don't run info for every refresh just to get the client name - only run on first time, and re-run on manual refresh

Don't clean the current files at the start of the refresh. This improves the perceived performance

Clean and populate is all done in one go so that running two refreshes doesn't lead to duplication

Updated tests - added an onRefreshStarted event to tell that a refresh is happening. Not ideal.

TODO - a test of multiple changelists having shelved files