add tests for initializing the set of resources - some tests todo
Need to investigate why the tests take this long - probably creating the resources via vscode API. Commiting to see how the CI handles this.

in order to sensibly do the configuration tests, the config accessing used in the model should ideally be refactored out so that it is possible to stub config settings.