Replace bottleneck with custom limiter. Refactor model config
Using bottleneck was causing about a half second delay for each command.
Created a CommandLimiter that performs the important part of the job (maxConcurrent) and removed the config items for bottleneck that were not used.

This reduces refresh time significantly

Created a config service for getting config settings from vscode

Created unit tests for items that don't require the vscode API